CLUE-453 fix: restore focus to table title on edit-mode exit
Add the `useRestoreFocusOnEditExit` hook alongside `EditableTitleButton`
and apply it to both `EditableTitleButton` callers. When the inline
title <input> unmounts on Enter / Escape, focus falls back to
`document.body` and the surrounding focus trap has no anchor for the
next Tab; the hook refocuses the display-mode title text, guarded so
a click that dismisses the editor doesn't yank focus from wherever
the user clicked.

The geometry / coordinate-grid case already had this fix inline in
`EditableTileTitle` (CLUE-503 upstream). Extracting the hook removes
the duplication and brings the table title to parity.
