+++ src/components/tiles/editable-tile-title.tsx
@@ -1,9 +1,9 @@
 import classNames from "classnames";
 import { observer } from "mobx-react";
-import React, { useContext, useEffect, useRef, useState } from "react";
+import React, { useContext, useRef, useState } from "react";
 import { useReadOnlyContext } from "../document/read-only-context";
 import { TileModelContext } from "../tiles/tile-api";
-import { EditableTitleButton } from "./editable-title-button";
+import { EditableTitleButton, useRestoreFocusOnEditExit } from "./editable-title-button";
 import { TileLabelInput } from "./tile-label-input";
 
 /** Callback to programmatically set title text. */
@@ -41,21 +41,7 @@ export const EditableTileTitle: React.FC<IProps> = observer(({
   const editingTitleRef = useRef(editingTitle);
   editingTitleRef.current = editingTitle;
 
-  // When edit mode exits via keyboard (Enter / Escape), focus falls back to
-  // body because the input unmounts; restore focus to the title text so the
-  // trap has a stable anchor for the next Tab. When edit mode exits because
-  // the user clicked elsewhere, leave focus on whatever they clicked.
-  const titleTextRef = useRef<HTMLDivElement>(null);
-  const prevIsEditingRef = useRef(false);
-  useEffect(() => {
-    if (prevIsEditingRef.current && !isEditing) {
-      const active = document.activeElement;
-      if (!active || active === document.body) {
-        titleTextRef.current?.focus();
-      }
-    }
-    prevIsEditingRef.current = isEditing;
-  }, [isEditing]);
+  const titleTextRef = useRestoreFocusOnEditExit(isEditing);
 
   const handleClick = () => {
     if (!readOnly && !isEditing) {
