+++ src/components/tiles/table/editable-table-title.tsx
@@ -3,7 +3,7 @@ import { observer } from "mobx-react";
 import React, { useState } from "react";
 import { ITileModel } from "../../../models/tiles/tile-model";
 import { verifyAlive } from "../../../utilities/mst-utils";
-import { EditableTitleButton } from "../editable-title-button";
+import { EditableTitleButton, useRestoreFocusOnEditExit } from "../editable-title-button";
 import { HeaderCellInput } from "./header-cell-input";
 
 interface IProps {
@@ -27,6 +27,9 @@ export const EditableTableTitle: React.FC<IProps> = observer(function EditableTa
   const title = model.computedTitle;
   const [isEditing, setIsEditing] = useState(false);
   const [editingTitle, setEditingTitle] = useState(title);
+
+  const titleTextRef = useRestoreFocusOnEditExit(isEditing);
+
   const handleClick = () => {
     if (!readOnly && !isEditing) {
       onBeginEdit?.();
@@ -60,7 +63,7 @@ export const EditableTableTitle: React.FC<IProps> = observer(function EditableTa
       {isEditing
         ? <HeaderCellInput style={style} value={editingTitle || ""}
             onKeyDown={handleKeyDown} onChange={setEditingTitle} onClose={handleClose} />
-        : <EditableTitleButton className="editable-table-title-text"
+        : <EditableTitleButton ref={titleTextRef} className="editable-table-title-text"
             title={title} readOnly={readOnly} onActivate={handleClick} />}
     </div>
   );
